+++ test/functional-tests/node-transformers/rename-properties-transformers/rename-properties-transformer/fixtures/property-definition-1.js
@@ -1,0 +1,7 @@
+const hawk = 'hawk';
+class Foo {
+    bar = 1;
+    static 'baz' = 2;
+    ['bark'] = 3;
+    [hawk] = 4;
+}