+++ test/functional-tests/javascript-obfuscator/fixtures/private-identifier.js
@@ -1,0 +1,7 @@
+class Foo {
+    #bar = 1;
+
+    method() {
+        this.#bar = 2;
+    }
+}