+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 v2.18.0
 ---
 * Added support of `es2022` features: private identifiers and class properties
+* Dropped support for `node@15`. 
+* Increased minimum supported `node` versions: `^12.22.0 || ^14.17.0 || >=16.0.0`
 
 v2.17.0
 ---